1.6.8源码

+++ erupt-core/src/main/java/xyz/erupt/core/view/Page.java
@@ -18,7 +18,7 @@ public class Page {
 
     private int pageSize;
 
-    private int totalPage;
+    private long totalPage;
 
     private Long total;
